[feat] 마트 상세 정보 및 프로덕트 디비 연결

+++ app/src/main/java/com/example/foodvillage/schema/Product.kt
@@ -4,8 +4,8 @@ data class Product(
     var storeName: String? = null,
     var productName: String? = null,
     var discountRate: Double? = null,
-    var fixedPrice: Int? = null,
+    var fixedPrice: Long? = null,
     var dibPeople: List<String>? = null,
-    var categoryNum: Int? = null,
+    var categoryNum: Long? = null,
     var imgUrl: String? = null
 )
